Add support for setuptools
For this we need to rename virt-v2v-wrapper naming to snake_case, because if
kebab-case is used for naming, there is no way to use the path to the module in
`setup.py`.

Due to that we need to adjust the imports in tests.  Because of the imports from
tests we need to make wrapper a module (I've no idea why everything else works
with the `__init__.py`).  And in order to make that available during the tests
we need to specify `usedevelop = True` in `tox.ini`.  And for the `usedevelop =
True` we need to have support for setuptools.  So this cannot be split into
smaller patches.

Also the install prefix needs to point somewhere python can import
from (otherwise the build fails), so the prefix needs to default to what python
tells us, hence the usage of `python-config --prefix`.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>