Added pre-transfer check for OSP volume, waiting for 'available'
Added a check before the OSP volume transfer begins, so that
we do not attempt to move the volume before it it is possible.
OSP volume has to be in the 'available' state before the transfer
begins. We are looping through TIMEOUT seconds (currently 300)
before giving up and erroring out.