Properties shouldn't be used unless necessary
The reason for properties to exist is to be able to change an API later on or do
a specific additional tasks in getters/setters.  Otherwise the code is just
slower and more bloated.  Properties will work after some clean-ups in the
`singleton.py` file, but they will not be needed, so the comments are not
relevant any more.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>