Fix rhv-verifypeer in a compatible way
The `rhv-verifypeer` option was only supplied if `data['insecure_connection']`
was `True`.  That means if someone wanted the connection to be secure it would
not specify any argument to virt-v2v.  Unfortunately, for some reasons,
rhv-upload's `rhv-verify` parameter defaults to `false` if unspecified.  So
there is no way of making sure the certificates are validated.

To mitigate this, always specify `rhv-verifypeer` so that we do not depend on
the defaults of virt-v2v's `rhv-upload`.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>